docs_site/requirements.txt
python-fasthtml
asyncstdlib